add separate component for RegisterProForm and login check
