+++ src/components/sign-up/register-pro-form/mock-data.js
@@ -1,0 +1,36 @@
+export const testArr = [
+	{
+		name: 'Test 1',
+		value: 'test1'
+	},{
+		name: 'Test 2',
+		value: 'test2'
+	}
+],
+iamArr = [
+	{
+		name: 'Individual',
+		value: 'individual'
+	},{
+		name: 'Company',
+		value: 'company'
+	}
+],
+minmaxArr = [
+	{
+		name: 'min',
+		value: 'min'
+	},{
+		name: 'max',
+		value: 'max'
+	}
+],
+monetaryUnitArr = [
+	{
+		name: 'Unit 1',
+		value: 'unit1'
+	},{
+		name: 'Unit 2',
+		value: 'unit2'
+	}
+]
