+++ src/routes/register-pro/style.scss
@@ -1,151 +1,2 @@
 
 @import './src/style/variables';
-
-$font-size: 18px;
-
-
-.registerPro{
-
-	.content{
-		width: 700px;
-		margin: 0 auto;
-
-		h2 {
-			text-align: center;
-			color: $red;
-			margin-top: 40px;
-		}
-
-		.applyArea{
-			margin-top: 30px;
-
-			.codeMsg, .timer {
-				font-size: 20px;
-				text-align: center;
-			}
-
-			.inputContainer{
-				width: 50%;
-				margin: 0 auto;
-			}
-
-			.verifyInput {
-				outline: none;
-				border: none;
-				border-bottom: 2px solid $lightGrey;
-				width: 50px;
-			}
-
-			.success {
-				font-size: 20px;
-				padding: 20px 0;
-				text-align: center;
-				margin-bottom: 20px;
-			
-			}
-		}
-		
-
-		.generalInfo{
-			margin-bottom: 50px;
-
-			& > div {
-				margin-bottom: 10px;
-			}
-		}
-
-		.generalField{
-			font-size: $font-size;
-			display: inline-block;
-
-			&:first-child{
-				width: 80px;
-			}
-		}
-
-		.radioLabel {
-			display: inline-block;
-			padding: 3px 0;
-			font-size: $font-size;
-			white-space: nowrap;
-			width: 80px;
-		}
-		
-		& > div {
-			margin-bottom: 15px;
-		}
-
-		button {
-			height: 35px;
-			width: 100%;
-			font-size: 18px;
-			background-color: $red;
-			border: none;
-			border-radius: 10px;
-			font-weight: bold;
-			cursor: pointer;
-		}
-	
-		.fieldContainer{
-			display: flex;
-			position: relative;
-			font-size: $font-size;
-			
-			label {
-				padding: 3px 0;
-				margin-right: 20px;
-				font-size: $font-size;
-				white-space: nowrap;
-				width: auto;
-			}
-
-			input[type=date] {
-				background: transparent;
-				width: 100%;
-				height: 100%;
-				outline: none;
-				color: $blue;
-				font-size: $font-size;
-				border: none;
-				border-bottom: 1px solid $lightGrey;
-				padding: 3px 0;
-			}
-	
-			.taElement {
-				width: 100%;
-				position: relative;
-				overflow: hidden;
-
-				textarea {
-					font-size: $font-size;
-					color: $blue;
-					resize: none;
-				  }
-				  textarea {
-					width: 100%;
-					box-sizing: border-box;
-					background-color: transparent;
-					
-					outline: none;
-					display: block;
-					line-height: 1.2;
-					border: none;
-					border-bottom: 2px solid $lightGrey;
-				  }
-				
-				  &:before {
-					position: absolute;
-					content: "_____________________________________________________________________________________________";
-					z-index: -1;
-					left: 0;
-					top: 0;
-					width: 100%;
-					color: $lightGrey;
-					line-height: 2;  
-				  }
-			  }
-			  
-			}
-	}
-	
-}
